common/requirements.txt
boto3
attrs >= 19.1.0
setuptools
